feat(view): extract navigation controls and add responsive font sizing
