+++ src/core/static/data.ts
@@ -18,8 +18,10 @@ export const staticData = {
       .href,
   },
   assets: {
-    surahFrame: new URL("../../data/assets/surah-frame.svg", import.meta.url)
-      .href,
+    surahFrame: new URL(
+      "../../data/assets/surah-frame-sides.svg",
+      import.meta.url,
+    ).href,
   },
 } as const;
